+++ src/react/index.ts
@@ -9,8 +9,6 @@ export {
 } from "./context/index.js";
 
 export * from "./hooks/index.js";
-// TODO: remove export with release 3.8
-export { SuspenseCache } from "./cache/index.js";
 
 export type { IDocumentDefinition } from "./parser/index.js";
 export { DocumentType, operationName, parser } from "./parser/index.js";